Creating a SiteStatsInformation singleton to store/access site related information in Stats related files.
